Add button benchmark
I also realized that the label benchmark wasn't testing the painting, since the buttons at some point will be placed outside the screen_rect, meaning it won't be painted.

This fixes it by benching the label in a child ui.

The label &str benchmark went from 483 ns to 535 ns with these changes.