Do not cache image data when it's been forgotten/cancelled while loading it
This is the same issue that was fixed for the http bytes loader in
239ade9a59692f23a6b1d779a9c2631cf774896a. However the mentioned fix was
wrong as it kept the loaded data in the cache even in the case where the
Pending cache entry was removed while loading (i.e. the image was forgotten).
Instead, use the same code as ab9f55ab0144462c1bc46a1fc49fe3b18f4d0fa7
(file loader) but for the ehttp and image loaders.