Removed unreachable code-path
This code is never reached as "GlowWinitRunning" is dropped earlier on
as part of the exiting process in response to the winit "CloseRequested"
event.
The winit "Destroyed" event only ever occurs after "CloseRequested".

The exception to this is if a window was not properly closed before the
event loop was excited previously.